api_tests/requirements.txt
requests>=2.24.0
python-ldap-test~=0.3.1
